Began work for texturing in RayMarchingTest
Not complete yet.

+++ test/RayMarchingTest/include/DeltaTime.hpp
@@ -7,46 +7,46 @@ class DeltaTime final
     DEFAULT_DESTRUCT(DeltaTime);
     DEFAULT_CM_PU(DeltaTime);
 private:
-    float _us;
-    float _ms;
-    float _s;
-    float _usSinceLastUpdate;
-    float _msSinceLastUpdate;
-    float _sSinceLastUpdate;
-    float _usSinceLaunch;
-    float _msSinceLaunch;
-    float _sSinceLaunch;
+    double _us;
+    double _ms;
+    double _s;
+    double _usSinceLastUpdate;
+    double _msSinceLastUpdate;
+    double _sSinceLastUpdate;
+    double _usSinceLaunch;
+    double _msSinceLaunch;
+    double _sSinceLaunch;
 public:
     DeltaTime() noexcept;
 
-    [[nodiscard]] float us() const noexcept { return _us; }
-    [[nodiscard]] float ms() const noexcept { return _ms; }
-    [[nodiscard]] float s() const noexcept { return _s; }
-    [[nodiscard]] float uSeconds() const noexcept { return _us; }
-    [[nodiscard]] float mSeconds() const noexcept { return _ms; }
-    [[nodiscard]] float seconds() const noexcept { return _s; }
-    [[nodiscard]] float microseconds() const noexcept { return _us; }
-    [[nodiscard]] float milliseconds() const noexcept { return _ms; }
+    [[nodiscard]] double us() const noexcept { return _us; }
+    [[nodiscard]] double ms() const noexcept { return _ms; }
+    [[nodiscard]] double s() const noexcept { return _s; }
+    [[nodiscard]] double uSeconds() const noexcept { return _us; }
+    [[nodiscard]] double mSeconds() const noexcept { return _ms; }
+    [[nodiscard]] double seconds() const noexcept { return _s; }
+    [[nodiscard]] double microseconds() const noexcept { return _us; }
+    [[nodiscard]] double milliseconds() const noexcept { return _ms; }
 
-    [[nodiscard]] float usSinceLastUpdate() const noexcept { return _usSinceLastUpdate; }
-    [[nodiscard]] float msSinceLastUpdate() const noexcept { return _msSinceLastUpdate; }
-    [[nodiscard]] float sSinceLastUpdate() const noexcept { return _sSinceLastUpdate; }
-    [[nodiscard]] float uSecondsSinceLastUpdate() const noexcept { return _usSinceLastUpdate; }
-    [[nodiscard]] float mSecondsSinceLastUpdate() const noexcept { return _msSinceLastUpdate; }
-    [[nodiscard]] float secondsSinceLastUpdate() const noexcept { return _sSinceLastUpdate; }
-    [[nodiscard]] float microsecondsSinceLastUpdate() const noexcept { return _usSinceLastUpdate; }
-    [[nodiscard]] float millisecondsSinceLastUpdate() const noexcept { return _msSinceLastUpdate; }
+    [[nodiscard]] double usSinceLastUpdate() const noexcept { return _usSinceLastUpdate; }
+    [[nodiscard]] double msSinceLastUpdate() const noexcept { return _msSinceLastUpdate; }
+    [[nodiscard]] double sSinceLastUpdate() const noexcept { return _sSinceLastUpdate; }
+    [[nodiscard]] double uSecondsSinceLastUpdate() const noexcept { return _usSinceLastUpdate; }
+    [[nodiscard]] double mSecondsSinceLastUpdate() const noexcept { return _msSinceLastUpdate; }
+    [[nodiscard]] double secondsSinceLastUpdate() const noexcept { return _sSinceLastUpdate; }
+    [[nodiscard]] double microsecondsSinceLastUpdate() const noexcept { return _usSinceLastUpdate; }
+    [[nodiscard]] double millisecondsSinceLastUpdate() const noexcept { return _msSinceLastUpdate; }
 
-    [[nodiscard]] float usSinceLaunch() const noexcept { return _usSinceLaunch; }
-    [[nodiscard]] float msSinceLaunch() const noexcept { return _msSinceLaunch; }
-    [[nodiscard]] float sSinceLaunch() const noexcept { return _sSinceLaunch; }
-    [[nodiscard]] float uSecondsSinceLaunch() const noexcept { return _usSinceLaunch; }
-    [[nodiscard]] float mSecondsSinceLaunch() const noexcept { return _msSinceLaunch; }
-    [[nodiscard]] float secondsSinceLaunch() const noexcept { return _sSinceLaunch; }
-    [[nodiscard]] float microsecondsSinceLaunch() const noexcept { return _usSinceLaunch; }
-    [[nodiscard]] float millisecondsSinceLaunch() const noexcept { return _msSinceLaunch; }
+    [[nodiscard]] double usSinceLaunch() const noexcept { return _usSinceLaunch; }
+    [[nodiscard]] double msSinceLaunch() const noexcept { return _msSinceLaunch; }
+    [[nodiscard]] double sSinceLaunch() const noexcept { return _sSinceLaunch; }
+    [[nodiscard]] double uSecondsSinceLaunch() const noexcept { return _usSinceLaunch; }
+    [[nodiscard]] double mSecondsSinceLaunch() const noexcept { return _msSinceLaunch; }
+    [[nodiscard]] double secondsSinceLaunch() const noexcept { return _sSinceLaunch; }
+    [[nodiscard]] double microsecondsSinceLaunch() const noexcept { return _usSinceLaunch; }
+    [[nodiscard]] double millisecondsSinceLaunch() const noexcept { return _msSinceLaunch; }
 
-    void setDeltaMicro(float us) noexcept;
+    void setDeltaMicro(double us) noexcept;
 
     void onUpdate() noexcept;
 };
